fix: move default mode contents to ./default
