+++ src/components/VariantSelector/default/markup.ts
@@ -1,5 +1,5 @@
 import { html } from "@/templating/html"
-import type { VariantSelector } from "./VariantSelector"
+import type { VariantSelector } from "../VariantSelector"
 import { ShopifyOption, ShopifyOptionValue, ShopifyProduct } from "@/shopify/graphql/types"
 
 export function generateVariantSelectorHTML(_element: VariantSelector, product: ShopifyProduct) {
